requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
jinja2==3.1.4
pydantic==2.9.1
pydantic-settings==2.5.2
httpx==0.27.0
websockets==12.0
sqlalchemy==2.0.32
alembic==1.13.2
hyperliquid-python-sdk==0.19.0
cryptography==43.0.1
python-dotenv==1.0.1
pytest==7.4.4
pytest-asyncio==0.21.1
itsdangerous==2.2.0
python-multipart==0.0.9
